refactor(cli): extract shared spinner helpers
approve-builds.ts had a copy of the getSpinner/getSilentSpinner shims
(makeSpinner) because it can't import them from prompts.ts: prompts.ts
already imports from approve-builds.ts, so reusing them directly would
create an import cycle.

Move getSpinner and getSilentSpinner into a shared utils/spinner.ts that
prompts.ts, approve-builds.ts, and the migrator all import from, dropping
the duplicate. No behavior change.
